requirements.txt
Django==5.1.7
djangorestframework==3.15.2
openai==1.65.4
python-dotenv==1.0.1
